requirements.txt
dateparser==1.2.0
Flask==3.0.3
flask_ngrok==0.0.25
Pillow==11.0.0
PyPDF2==3.0.1
python_dateutil==2.8.2
pythonProject==1.0.0
spacy==3.7.4
torch==1.12.1
transformers==4.36.2
Werkzeug==3.0.6
word2number==1.1
